Begin creating Java equivalents of some OCL functionality:
IntegrationIntegrationModelHelper.java + bug fix

This is because we occasionally need more output flexibility than we get
in OCL. This only implements in Java the things we need now. More will
be converted later as needed.

Also - a bug fix: closure will not include the starting point if its a
different type, so all IMs must include self beyond containedIMs