Clean up references between OCL files.
Created 'all-....ocl' files to directly load invariant files.
This replaces false dependencies betweeen ocl files that appeared when
splitting between invariants and extension methods. When both are in a
single file, that creates dependencies that automatically load all the
invariants. When the extensions are split out, then the dependencies
change to those extensions and invariants won't be automatically loaded.

It would be possible to leave those invariants inclusions but this would
create a false dependency. Instead, all invariants now can be loaded
directly using one of the 'all-...ocl' files.